Handle nilable arrays
