+++ ruby/optify/lib/optify_ruby/options_metadata.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+# typed: true
+
+require 'sorbet-runtime'
+
+require_relative './base_config'
+
+module Optify
+  # Information about a feature.
+  class OptionsMetadata < BaseConfig
+    sig { returns(T.nilable(T::Array[String])) }
+    attr_reader :aliases
+
+    sig { returns(T.untyped) }
+    attr_reader :details
+
+    sig { returns(String) }
+    attr_reader :name
+
+    sig { returns(T.nilable(String)) }
+    attr_reader :owners
+  end
+end
