+++ ruby/optify/optify.gemspec
@@ -38,11 +38,11 @@ Gem::Specification.new do |spec|
   # needed until rubygems supports Rust support is out of beta
   spec.add_dependency 'rb_sys', '~> 0.9.109'
 
-  spec.add_dependency 'sorbet-runtime', '~> 0.5.11851'
+  spec.add_dependency 'sorbet-runtime', '~> 0.5.11911'
 
   spec.add_development_dependency 'rake-compiler', '~> 1.2.9'
   spec.add_development_dependency 'rbs', '~> 3.8.1'
-  spec.add_development_dependency 'sorbet', '~> 0.5.11851'
+  spec.add_development_dependency 'sorbet', '~> 0.5.11911'
   spec.add_development_dependency 'tapioca', '~> 0.16.11'
 
   # Tests
