+++ ruby/optify/sorbet/rbi/gems/spoom@1.5.4.rbi
@@ -1027,7 +1027,7 @@ class Spoom::Coverage::D3::ColorPalette < ::T::Struct
   prop :strong, ::String
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -1367,7 +1367,7 @@ class Spoom::Coverage::Snapshot < ::T::Struct
     sig { params(obj: T::Hash[::String, T.untyped]).returns(::Spoom::Coverage::Snapshot) }
     def from_obj(obj); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -1503,7 +1503,7 @@ class Spoom::Deadcode::Definition < ::T::Struct
   def to_json(*args); end
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2391,7 +2391,7 @@ class Spoom::Deadcode::Send < ::T::Struct
   def each_arg_assoc(&block); end
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2411,7 +2411,7 @@ class Spoom::ExecResult < ::T::Struct
   def to_s; end
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2609,7 +2609,7 @@ class Spoom::FileTree::Node < ::T::Struct
   def path; end
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2675,7 +2675,7 @@ class Spoom::Git::Commit < ::T::Struct
   def timestamp; end
 
   class << self
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
 
     # Parse a line formatted as `%h %at` into a `Commit`
@@ -2787,7 +2787,7 @@ class Spoom::LSP::Diagnostic < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::Diagnostic) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2820,7 +2820,7 @@ class Spoom::LSP::DocumentSymbol < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::DocumentSymbol) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2878,7 +2878,7 @@ class Spoom::LSP::Hover < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::Hover) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2903,7 +2903,7 @@ class Spoom::LSP::Location < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::Location) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -2966,7 +2966,7 @@ class Spoom::LSP::Position < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::Position) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -3004,7 +3004,7 @@ class Spoom::LSP::Range < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::Range) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -3070,7 +3070,7 @@ class Spoom::LSP::SignatureHelp < ::T::Struct
     sig { params(json: T::Hash[T.untyped, T.untyped]).returns(::Spoom::LSP::SignatureHelp) }
     def from_json(json); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
   end
 end
@@ -3496,7 +3496,7 @@ class Spoom::Model::Reference < ::T::Struct
     sig { params(name: ::String, location: ::Spoom::Location).returns(::Spoom::Model::Reference) }
     def constant(name, location); end
 
-    # source://sorbet-runtime/0.5.11796/lib/types/struct.rb#13
+    # source://sorbet-runtime/0.5.11911/lib/types/struct.rb#13
     def inherited(s); end
 
     # source://spoom//lib/spoom/model/reference.rb#29
@@ -4348,36 +4348,36 @@ end
 
 # From https://github.com/Shopify/ruby-lsp/blob/9154bfc6ef/lib/ruby_lsp/document.rb#L127
 #
-# source://spoom//lib/spoom/sorbet/sigs.rb#145
+# source://spoom//lib/spoom/sorbet/sigs.rb#147
 class Spoom::Sorbet::Sigs::Scanner
-  # source://spoom//lib/spoom/sorbet/sigs.rb#151
+  # source://spoom//lib/spoom/sorbet/sigs.rb#153
   sig { params(source: ::String).void }
   def initialize(source); end
 
   # Finds the character index inside the source string for a given line and column
   #
-  # source://spoom//lib/spoom/sorbet/sigs.rb#159
+  # source://spoom//lib/spoom/sorbet/sigs.rb#161
   sig { params(line: ::Integer, character: ::Integer).returns(::Integer) }
   def find_char_position(line, character); end
 end
 
-# source://spoom//lib/spoom/sorbet/sigs.rb#148
+# source://spoom//lib/spoom/sorbet/sigs.rb#150
 Spoom::Sorbet::Sigs::Scanner::LINE_BREAK = T.let(T.unsafe(nil), Integer)
 
-# source://spoom//lib/spoom/sorbet/sigs.rb#83
+# source://spoom//lib/spoom/sorbet/sigs.rb#85
 class Spoom::Sorbet::Sigs::SigTranslator
   class << self
-    # source://spoom//lib/spoom/sorbet/sigs.rb#88
+    # source://spoom//lib/spoom/sorbet/sigs.rb#90
     sig { params(sig: ::RBI::Sig, node: T.any(::RBI::Attr, ::RBI::Method)).returns(::String) }
     def translate(sig, node); end
 
     private
 
-    # source://spoom//lib/spoom/sorbet/sigs.rb#135
+    # source://spoom//lib/spoom/sorbet/sigs.rb#137
     sig { params(sig: ::RBI::Sig, node: ::RBI::Attr).returns(::String) }
     def translate_attr_sig(sig, node); end
 
-    # source://spoom//lib/spoom/sorbet/sigs.rb#100
+    # source://spoom//lib/spoom/sorbet/sigs.rb#102
     sig { params(sig: ::RBI::Sig, node: ::RBI::Method).returns(::String) }
     def translate_method_sig(sig, node); end
   end
